% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeLambda.R
\name{ComputeLambda}
\alias{ComputeLambda}
\title{L1 norm of a projection}
\usage{
ComputeLambda(uFT, pResidual)
}
\arguments{
\item{uFT}{A vector of 0s and 1s of lenght nc: (-1)^uFT is a unit vector in Linf  norm}

\item{pResidual}{A matrix with nc columns}
}
\value{
L1 norm of the pResidual x (-1)^uFT
}
\description{
L1 norm of a projection
}
\details{
This function is for internal usage only.
}
\examples{
ComputeLambda(uFT=c(FALSE,TRUE,FALSE),pResidual=matrix(1:15,nr=5,nc=3))

}
