% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAllBinaries.R
\name{CreateAllBinaries}
\alias{CreateAllBinaries}
\title{Create all 2^n n-tuples of ab[1] and ab[2]}
\usage{
CreateAllBinaries(n = 1, ab = c(0, 1))
}
\arguments{
\item{n}{An integer > 0}

\item{ab}{A vector of length 2}
}
\value{
a n x 2^n matrix
}
\description{
Create all 2^n n-tuples of ab[1] and ab[2]
}
\details{
This function is for internal usage only.
}
\examples{
 CreateAllBinaries(3,c(0,1))

}
