% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JitterPosition.R
\name{JitterPosition}
\alias{JitterPosition}
\title{Moves points that are close together}
\usage{
JitterPosition(x, y, jitterDefault = 0.01, Near = 0.045)
}
\arguments{
\item{x}{The x-coordinates of the points}

\item{y}{The y-coordinates of the points}

\item{jitterDefault}{A positive number controlling the maximum jitter}

\item{Near}{A positive number controlling the definition of "near"}
}
\value{
A list giving the new x and y coordinates, and, for each point, its number neighbours and a suggested relative size for the plotting symbol
}
\description{
Moves points that are close together
}
\details{
This function is for internal usage only.
}
\examples{
JitterPosition(c(1:5,2,2,4,4,4),c(1:5,2,2,4,4,4))

}
