% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListToObjects.R
\name{ListToObjects}
\alias{ListToObjects}
\title{Extract objects from a list}
\usage{
ListToObjects(L, envir = .GlobalEnv)
}
\arguments{
\item{L}{A list}

\item{envir}{The environment into which variables are created}
}
\value{
No return
}
\description{
Extract objects from a list to the global environment
}
\details{
This function is for internal usage only.
}
\examples{
ListToObjects(list(x=5,A="Hello",M=matrix(1:8,nr=2)),envir=.GlobalEnv)
}
