% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tca.r
\name{print.tca}
\alias{print.tca}
\title{Print result of Taxicab Analysis in easily readable format}
\usage{
\method{print}{tca}(x, ...)
}
\arguments{
\item{x}{A tca tcaObject produced by the function tca}

\item{...}{Unused}
}
\value{
An invisible list containing formated outputs
}
\description{
Print result of Taxicab Analysis in easily readable format
}
\examples{
print(tca(rodent))
}
