% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rodent}
\alias{rodent}
\title{Rodent species abundance}
\format{A data frame of counts with 28 rows and 9 columns
\itemize{
  \item Columns: Rodent species
  \item Rows: Location
}}
\source{
Bolger et al. 1997, Response of rodents to habitat fragmentation in coastal Southern California, Ecological Applications 7, 552-563 (modified) via \url{https://www.zoology.ubc.ca/~bio501/R/workshops/workshops-multivariate-methods/}
}
\usage{
data(rodent)
}
\description{
Counts of rodents by species and site
}
\examples{
tca(rodent,nAxes=4)
}
\keyword{datasets}
