% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveTCA.r
\name{saveTCA}
\alias{saveTCA}
\title{Save tca results to a folder}
\usage{
saveTCA(tcaObject, path, folder = NULL, what = c("report", "csv", "plot",
  "dataMatrix", "tcaObject"), plotAxes = matrix((1:2), nrow = 1, ncol = 2,
  byrow = T), graphicDevice = c("pdf", "postscript", "xfig", "bitmap",
  "pictex", "cairo_pdf", "cairo_ps", "svg", "png", "jpeg", "bmp", "tiff"),
  csvFormat = c("csv", "csv2"))
}
\arguments{
\item{tcaObject}{a tca-class object created by tca}

\item{path}{Location of the folder}

\item{folder}{Name of the folder}

\item{what}{What to save: all items specified will be saved}

\item{plotAxes}{A k x 2 matrix giving pairs of axes to plot and save}

\item{graphicDevice}{Format(s) of plots saved.  Plots can be saved in more than one format}

\item{csvFormat}{Format of csv files (North American or European)}
}
\value{
Figure
}
\description{
Save tca results to a folder
}
\examples{
saveTCA(tca(rodent),path=tempdir())
}
