% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchCrissCross.R
\name{SearchCrissCross}
\alias{SearchCrissCross}
\title{Search a taxicab principal component using the criss-cross algorithm}
\usage{
SearchCrissCross(pResidual, iterationMax = 20)
}
\arguments{
\item{pResidual}{A matrix of of non-negative numbers}

\item{iterationMax}{Maximum number of iterations}
}
\value{
A list: L1Max = maximum L1 norm; uMax = Linf unit vector giving the maximum L1 norm
}
\description{
Search a taxicab principal component using the criss-cross algorithm
}
\details{
This function is for internal usage only.

The vector uMax is normalize to uMax[1] = 1
}
\examples{
SearchCrissCross(matrix(-3:8,nrow=4,ncol=3))

}
