% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchExhaustive.R
\name{SearchExhaustive}
\alias{SearchExhaustive}
\title{Search a taxicab principal component via exhaustive search}
\usage{
SearchExhaustive(pResidual)
}
\arguments{
\item{pResidual}{A matrix of of non-negative numbers}
}
\value{
A list: L1Max = maximum L1 norm; uMax = Linf unit vector giving the maximum L1 norm
}
\description{
Search a taxicab principal component via exhaustive search
}
\details{
This function is for internal usage only.

The vector uMax is normalize to uMax[1] = 1
}
\examples{
SearchExhaustive(matrix(-3:8,nrow=4,ncol=3))

}
