% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchGeneticAlgoritm.R
\name{SearchGeneticAlgoritm}
\alias{SearchGeneticAlgoritm}
\title{Search a taxicab principal component using the genetic algorithm}
\usage{
SearchGeneticAlgoritm(pResidual)
}
\arguments{
\item{pResidual}{A matrix of of non-negative numbers}
}
\value{
A list: L1Max = maximum L1 norm; uMax = Linf unit vector giving the maximum L1 norm
}
\description{
Search a taxicab principal component using the genetic algorithm
}
\details{
This function is for internal usage only.

The vector uMax is normalize to uMax[1] = 1
}
\examples{
SearchGeneticAlgoritm(matrix(-3:8,nrow=4,ncol=3))

}
