% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tca.r
\name{summary.tca}
\alias{summary.tca}
\title{Summary of the Taxicab analysis}
\usage{
\method{summary}{tca}(object, ...)
}
\arguments{
\item{object}{A tca tcaObject produced by the function tca}

\item{...}{Unused}
}
\value{
A list
}
\description{
Summary of the Taxicab analysis
}
\details{
Shows the unstandardized dispersion values
}
\examples{
summary(tca(rodent))
}
